12_base_tool/zone_system/15_bt_journal/requirements.txt
# Epoch Backtest Journal - Dependencies

# Database
psycopg2-binary>=2.9.9

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Statistical Analysis
scipy>=1.10.0

# PDF Generation & Charts
matplotlib>=3.7.0

# Excel Export
openpyxl>=3.1.0

# Optional: Data validation
# pydantic>=2.0.0
